+++ components/client/src/main/java/com/hotels/styx/client/retry/RetryNTimes.java
@@ -18,7 +18,7 @@ package com.hotels.styx.client.retry;
 import com.hotels.styx.api.client.RemoteHost;
 import com.hotels.styx.api.client.loadbalancing.spi.LoadBalancer;
 import com.hotels.styx.api.client.retrypolicy.spi.RetryPolicy;
-import com.hotels.styx.api.netty.exceptions.IsRetryableException;
+import com.hotels.styx.api.exceptions.IsRetryableException;
 
 import java.util.Optional;
