+++ components/client/src/test/integration/scala/com/hotels/styx/client/HttpClientSpec.scala
@@ -22,8 +22,8 @@ import com.hotels.styx.api.HttpRequest.get
 import com.hotels.styx.api.client.Origin._
 import com.hotels.styx.api.client.loadbalancing.spi.LoadBalancer
 import com.hotels.styx.api.client.{ActiveOrigins, Origin}
+import com.hotels.styx.api.exceptions.ResponseTimeoutException
 import com.hotels.styx.api.messages.HttpResponseStatus.OK
-import com.hotels.styx.api.netty.exceptions.ResponseTimeoutException
 import com.hotels.styx.api.service.BackendService
 import com.hotels.styx.client.OriginsInventory.newOriginsInventoryBuilder
 import com.hotels.styx.client.StyxHttpClient._
