+++ components/client/src/test/unit/java/com/hotels/styx/client/StyxHttpClientTest.java
@@ -28,8 +28,8 @@ import com.hotels.styx.api.client.loadbalancing.spi.LoadBalancer;
 import com.hotels.styx.api.client.retrypolicy.spi.RetryPolicy;
 import com.hotels.styx.api.metrics.MetricRegistry;
 import com.hotels.styx.api.metrics.codahale.CodaHaleMetricRegistry;
-import com.hotels.styx.api.netty.exceptions.NoAvailableHostsException;
-import com.hotels.styx.api.netty.exceptions.OriginUnreachableException;
+import com.hotels.styx.api.exceptions.NoAvailableHostsException;
+import com.hotels.styx.api.exceptions.OriginUnreachableException;
 import com.hotels.styx.api.service.BackendService;
 import com.hotels.styx.api.service.StickySessionConfig;
 import org.hamcrest.Matchers;
