+++ components/client/src/test/unit/java/com/hotels/styx/client/TransportTest.java
@@ -22,7 +22,7 @@ import com.hotels.styx.api.Id;
 import com.hotels.styx.api.StyxObservable;
 import com.hotels.styx.api.client.Connection;
 import com.hotels.styx.api.client.ConnectionPool;
-import com.hotels.styx.api.netty.exceptions.NoAvailableHostsException;
+import com.hotels.styx.api.exceptions.NoAvailableHostsException;
 import io.netty.buffer.ByteBuf;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
