+++ components/client/src/test/unit/java/com/hotels/styx/client/connectionpool/SimpleConnectionPoolTest.java
@@ -20,7 +20,7 @@ import com.hotels.styx.api.Id;
 import com.hotels.styx.api.client.Connection;
 import com.hotels.styx.api.client.ConnectionPool;
 import com.hotels.styx.api.client.Origin;
-import com.hotels.styx.api.netty.exceptions.TransportException;
+import com.hotels.styx.api.exceptions.TransportException;
 import com.hotels.styx.api.service.ConnectionPoolSettings;
 import com.hotels.styx.client.connectionpool.stubs.StubConnectionFactory;
 import com.hotels.styx.client.connectionpool.stubs.StubConnectionFactory.StubConnection;
