+++ components/client/src/test/unit/java/com/hotels/styx/client/netty/connectionpool/NettyToStyxResponsePropagatorTest.java
@@ -18,8 +18,8 @@ package com.hotels.styx.client.netty.connectionpool;
 import com.google.common.base.Throwables;
 import com.hotels.styx.api.HttpResponse;
 import com.hotels.styx.api.client.Origin;
-import com.hotels.styx.api.netty.exceptions.ResponseTimeoutException;
-import com.hotels.styx.api.netty.exceptions.TransportLostException;
+import com.hotels.styx.api.exceptions.ResponseTimeoutException;
+import com.hotels.styx.api.exceptions.TransportLostException;
 import com.hotels.styx.client.BadHttpResponseException;
 import com.hotels.styx.client.StyxClientException;
 import io.netty.buffer.ByteBuf;
