+++ components/common/src/main/java/com/hotels/styx/api/exceptions/IsDeadConnectionException.java
@@ -13,10 +13,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.netty.exceptions;
+package com.hotels.styx.api.exceptions;
 
 /**
- * A marker interface for exceptions that indicates that an exception was caused by a timeout.
+ * A marker interface for exceptions that indicate that a connection is dead.
  */
-public interface IsTimeoutException {
+public interface IsDeadConnectionException {
 }
