+++ components/common/src/main/java/com/hotels/styx/api/exceptions/OriginUnreachableException.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.netty.exceptions;
+package com.hotels.styx.api.exceptions;
 
 
 import com.hotels.styx.api.client.Origin;
